Import BS & Mesh Collider for Island
Files getting updated as part of an import of unity not "recognizing the editor version used."

It broke the shader due to unplugging the URP settings from the Project's Graphic Settings, but it was an easy fix.

Oh, and I also added a mesh collider to the new island. Missed that, somehow.