PineTree Import
Imports a new pine tree model. The old pinetree model has been renamed to "OldTree". A prefab also exists for the model using a toon shader, and two materials exist, one URP and One Toon.

NOTE: You may need to create an additional material for the tree to have it usable with the paint tree functionality.